requirements.txt
requests~=2.32.4
packaging~=26.0
ruamel.yaml~=0.19.1